done main fuctions

+++ status/views.py
@@ -3,6 +3,7 @@ from django.contrib.auth.decorators import login_required
 from status.models import Status
 from status.forms import StatusForm
 from django.contrib import messages
+from tasks.models import Task
 
 @login_required
 def status_task(request):
@@ -34,9 +35,13 @@ def edit_status(request, status_id):
     return render(request, 'edit_status.html', {'status': status})
 @login_required
 def delete_status(request, status_id):
-    status = Status.objects.get(pk=status_id)
+    status = get_object_or_404(Status, id=status_id)
     if request.method == 'POST':
-        status.delete()
+        if not Task.objects.filter(status=status).exists(): # Проверяем, используется ли статус в задачах
+            status.delete()
+            messages.success(request, 'Статус успешно удалён!')
+        else:
+            messages.error(request, 'Статус не может быть удалён, так как он используется в задаче.')
         return redirect('status_task')
 
     return render(request, 'status/confirm_delete.html', {'status': status})
